split Tools screen components
