+++ src/components/tools/constants.ts
@@ -1,0 +1,118 @@
+import type { Category, ToolCategory } from "./types";
+
+export const categoryColors: Record<ToolCategory, string> = {
+  Utility:  "bg-amber-500/10 text-amber-400 border-amber-500/20",
+  Text:     "bg-sky-500/10 text-sky-400 border-sky-500/20",
+  Data:     "bg-violet-500/10 text-violet-400 border-violet-500/20",
+  Media:    "bg-pink-500/10 text-pink-400 border-pink-500/20",
+  Network:  "bg-orange-500/10 text-orange-400 border-orange-500/20",
+  Security: "bg-red-500/10 text-red-400 border-red-500/20",
+  Storage:  "bg-emerald-500/10 text-emerald-400 border-emerald-500/20",
+  Platform: "bg-primary/10 text-primary border-primary/20",
+  Social:   "bg-blue-500/10 text-blue-400 border-blue-500/20",
+  Commerce: "bg-teal-500/10 text-teal-400 border-teal-500/20",
+};
+
+export const categoryIconBg: Record<ToolCategory, string> = {
+  Utility:  "bg-amber-500/10 text-amber-400",
+  Text:     "bg-sky-500/10 text-sky-400",
+  Data:     "bg-violet-500/10 text-violet-400",
+  Media:    "bg-pink-500/10 text-pink-400",
+  Network:  "bg-orange-500/10 text-orange-400",
+  Security: "bg-red-500/10 text-red-400",
+  Storage:  "bg-emerald-500/10 text-emerald-400",
+  Platform: "bg-primary/10 text-primary",
+  Social:   "bg-blue-500/10 text-blue-400",
+  Commerce: "bg-teal-500/10 text-teal-400",
+};
+
+export const PLATFORM_CONNECTOR_SLUGS: Record<string, string> = {
+  "Telegram":       "telegram",
+  "Slack":          "slack",
+  "Discord":        "discord",
+  "Reddit":         "reddit",
+  "Bluesky":        "bluesky",
+  "Mastodon":       "mastodon",
+  "LINE":           "line",
+  "Figma":          "figma",
+  "Shopify":        "shopify",
+  "Amazon":         "amazon",
+  "Xero":           "xero",
+  "eBay":           "ebay",
+  "Etsy":           "etsy",
+  "Stripe":         "stripe",
+  "PayPal":         "paypal",
+  "Square":         "square",
+  "QuickBooks":     "quickbooks",
+  "Plaid":          "plaid",
+  "WooCommerce":    "woocommerce",
+  "Passport":       "keychain",
+  "C-Suite Analyze":"csuite",
+  "Notion":         "notion",
+  "Vercel":         "vercel",
+  "Resend":         "resend",
+  "Email":          "email",
+  "Monica":         "monica",
+  "Clockify":       "clockify",
+  "Toggl":          "toggl",
+  "Twilio":         "twilio",
+  "Pushover":       "pushover",
+  "WhatsApp":       "whatsapp",
+  "YouTube":        "youtube",
+  "Spotify":        "spotify",
+  "ElevenLabs":     "elevenlabs",
+  "Replicate":      "replicate",
+  "Stability AI":   "stability",
+  "OpenAI":         "openai",
+  "Anthropic":      "anthropic",
+  "Higgsfield":     "higgsfield",
+  "HeyGen":         "heygen",
+  "Runway":         "runway",
+  "Pika":           "pika",
+  "Kling AI":       "kling",
+  "PagerDuty":      "pagerduty",
+  "CircleCI":       "circleci",
+  "Segment":        "segment",
+  "Postmark":       "postmark",
+  "Gumroad":        "gumroad",
+  "Together AI":    "togetherai",
+  "Asana":          "asana",
+  "Monday":         "monday",
+  "Calendly":       "calendly",
+  "Pinterest":      "pinterest",
+  "TikTok":         "tiktok",
+  "Steam":          "steam",
+  "Mailchimp":      "mailchimp",
+  "SendGrid":       "sendgrid",
+  "Datadog":        "datadog",
+  "Groq":           "groq",
+  "AssemblyAI":     "assemblyai",
+  "Pinecone":       "pinecone",
+  "Mixpanel":       "mixpanel",
+};
+
+export const PLATFORM_CONNECTOR_NAMES = new Set(Object.keys(PLATFORM_CONNECTOR_SLUGS));
+
+export const categories: Category[] = [
+  "All", "Local", "Platform",
+  "Utility", "Text", "Data", "Media", "Network", "Security", "Storage", "Social", "Commerce",
+];
+
+export const NO_API_KEY_TOOLS = new Set([
+  "Random", "UUID", "Cron", "Timestamp", "QR Code", "Units", "Weather", "Currency", "Notify",
+  "Transform", "Regex", "Diff", "Encode", "Readability", "Markdown",
+  "Hash", "DNS", "Headers", "Ping", "Whois", "Sitemap", "RSS",
+  "Calculator", "Unit Converter", "Datetime", "Text Tool", "Color Convert", "Random Generator", "Meal Planner",
+  "BoardGameGeek", "ESPN", "Sleeper", "Have I Been Pwned", "NVD",
+  "Open Food Facts", "Deezer",
+  "USGS", "OpenAQ", "eBird",
+  "Trove", "TAB", "The Lott", "Toilets",
+]);
+
+export const QUICK_LINKS: { label: string; category: Category }[] = [
+  { label: "Weather", category: "Utility" },
+  { label: "Security", category: "Security" },
+  { label: "Gaming", category: "Data" },
+  { label: "Social", category: "Social" },
+  { label: "Australia", category: "Data" },
+];
